requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
cffi==2.0.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.2
python-dotenv==1.2.1
tonsdk==1.0.15
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
